backend/requirements.txt
google-genai>=1.0.0
google-adk>=1.0.0
google-cloud-firestore>=2.19.0
google-cloud-aiplatform>=1.71.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
python-dotenv>=1.0.0
mcp>=1.0.0
